Move `fun aRoomSummaryDetails()` to a regular Preview Provider
